+++ paddle/fluid/eager/tests/CMakeLists.txt
@@ -1,3 +1,2 @@
 add_subdirectory(data_structure_tests)
-add_subdirectory(task_tests)
 add_subdirectory(performance_tests)
