+++ test/cpp/eager/task_tests/grad_test.cc
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/eager/backward.h"
-
 #include <sstream>
 
 #include "glog/logging.h"
@@ -23,8 +21,9 @@
 #include "paddle/fluid/eager/api/generated/eager_generated/backwards/scale_node.h"
 #include "paddle/fluid/eager/api/utils/tensor_utils.h"
 #include "paddle/fluid/eager/autograd_meta.h"
+#include "paddle/fluid/eager/backward.h"
 #include "paddle/fluid/eager/grad_node_info.h"
-#include "paddle/fluid/eager/tests/test_utils.h"
+#include "paddle/fluid/eager/test_utils.h"
 #include "paddle/phi/core/dense_tensor.h"
 #include "paddle/phi/core/kernel_registry.h"
 #include "paddle/phi/core/tensor_meta.h"
@@ -34,15 +33,15 @@ PD_DECLARE_KERNEL(add, CPU, ALL_LAYOUT);
 
 namespace egr {
 
-TEST(Backward, SingleNodeEmptyGrad) {
+TEST(Grad, SingleNodeEmptyGrad) {
   // Prepare Device Contexts
   eager_test::InitEnv(paddle::platform::CPUPlace());
 
   // Prepare Inputs
   paddle::framework::DDim ddim = phi::make_ddim({4, 16, 16, 32});
 
-  // Create Target Tensor
-  paddle::Tensor target_tensor =
+  // Create Target Tensor (output)
+  paddle::Tensor output_tensor =
       egr_utils_api::CreateTensorWithValue(ddim,
                                            paddle::platform::CPUPlace(),
                                            phi::DataType::FLOAT32,
@@ -50,7 +49,15 @@ TEST(Backward, SingleNodeEmptyGrad) {
                                            1.0 /*value*/,
                                            false /*is_leaf*/);
 
-  paddle::Tensor leaf_tensor;
+  // Create input tensor
+  const paddle::Tensor leaf_tensor =
+      egr_utils_api::CreateTensorWithValue(ddim,
+                                           paddle::platform::CPUPlace(),
+                                           phi::DataType::FLOAT32,
+                                           phi::DataLayout::NCHW,
+                                           1.0 /*value*/,
+                                           true /*is_leaf*/);
+
   {
     // Create Scale Node
     auto node0_ptr = std::make_shared<GradNodeScale>(1, 1);
@@ -58,34 +65,41 @@ TEST(Backward, SingleNodeEmptyGrad) {
 
     // Set grad in/out meta
     node0_ptr->SetDefaultGradInOutMeta();
-    AutogradMeta* auto_grad_meta = EagerUtils::autograd_meta(&target_tensor);
+
+    // Output_tensor set GradNode、OutRank、StopGradient propertis
+    AutogradMeta* auto_grad_meta = EagerUtils::autograd_meta(&output_tensor);
     auto_grad_meta->SetGradNode(
         std::dynamic_pointer_cast<GradNodeBase>(node0_ptr));
     auto_grad_meta->SetSingleOutRankWithSlot(0, 0);
     auto_grad_meta->SetStopGradient(false);
 
-    AutogradMeta* auto_grad_meta1 = EagerUtils::autograd_meta(&leaf_tensor);
+    // Get autograd_meta from input tensor
+    AutogradMeta* auto_grad_meta1 =
+        EagerUtils::unsafe_autograd_meta(leaf_tensor);
 
     // Connect Tensor and AccumulationNode via AutoGradMeta
     auto acc_node_ptr =
         std::make_shared<egr::GradNodeAccumulation>(auto_grad_meta1);
 
+    // input tensor set GradNode、OutRank、StopGradient propertis
     auto_grad_meta1->SetGradNode(
         std::dynamic_pointer_cast<GradNodeBase>(acc_node_ptr));
     auto_grad_meta1->SetSingleOutRankWithSlot(0, 0);
     auto_grad_meta1->SetStopGradient(false);
 
-    node0_ptr->SetGradOutMeta({leaf_tensor}, 0);
+    // grad_node Add Edges
+    std::vector<egr::AutogradMeta*> res = {auto_grad_meta1};
+    node0_ptr->SetGradOutMeta(leaf_tensor, 0);
   }
-  std::vector<paddle::Tensor> outs = {target_tensor};
-  // Run Backward
-  Backward(outs, {});
+  std::vector<paddle::Tensor> outs = {output_tensor};
 
+  // Run Grad
+  auto result = Grad(outs, {leaf_tensor}, {});
   // Check Output Value
-  eager_test::CompareGradTensorWithValue<float>(leaf_tensor, 5.0);
+  eager_test::CompareTensorWithValue<float>(result[0], 5.0);
 }
 
-TEST(Backward, SingleNodeCustomGrad) {
+TEST(Grad, SingleNodeCustomGrad) {
   // Prepare Device Contexts
   eager_test::InitEnv(paddle::platform::CPUPlace());
 
@@ -114,7 +128,14 @@ TEST(Backward, SingleNodeCustomGrad) {
                                            false /*is_leaf*/);
   grad_tensors.emplace_back(std::move(grad_tensor));
 
-  paddle::Tensor leaf_tensor;
+  paddle::Tensor leaf_tensor =
+      egr_utils_api::CreateTensorWithValue(ddim,
+                                           paddle::platform::CPUPlace(),
+                                           phi::DataType::FLOAT32,
+                                           phi::DataLayout::NCHW,
+                                           1.0 /*value*/,
+                                           true /*is_leaf*/);
+
   {
     // Create Scale Node
     auto node0_ptr = std::make_shared<GradNodeScale>(1, 1);
@@ -140,14 +161,14 @@ TEST(Backward, SingleNodeCustomGrad) {
         std::dynamic_pointer_cast<GradNodeBase>(acc_node_ptr));
     auto_grad_meta1->SetSingleOutRankWithSlot(0, 0);
     auto_grad_meta1->SetStopGradient(false);
-    node0_ptr->SetGradOutMeta({leaf_tensor}, 0);
+    std::vector<egr::AutogradMeta*> res = {auto_grad_meta1};
+    node0_ptr->SetGradOutMeta(leaf_tensor, 0);
   }
 
-  // Run Backward
-  Backward(target_tensors, grad_tensors);
+  auto result = Grad(target_tensors, {leaf_tensor}, grad_tensors);
 
   // Check Output Value
-  eager_test::CompareGradTensorWithValue<float>(leaf_tensor, 50.0);
+  eager_test::CompareTensorWithValue<float>(result[0], 50.0);
 }
 
 /*
@@ -155,13 +176,13 @@ Node1
   |
 Node0
   |
- inp0
+ { } // empty grad tensor
 */
-TEST(Backward, LinearNodes) {
+TEST(Grad, LinearNodes) {
   // Prepare Device Contexts
   eager_test::InitEnv(paddle::platform::CPUPlace());
 
-  // Prepare Inputs
+  // Prepare Target Tensor
   std::vector<paddle::Tensor> target_tensors;
   paddle::framework::DDim ddim = phi::make_ddim({4, 16, 16, 32});
 
@@ -175,7 +196,13 @@ TEST(Backward, LinearNodes) {
                                            false /*is_leaf*/);
   target_tensors.emplace_back(std::move(tensor));
 
-  paddle::Tensor leaf_tensor;
+  paddle::Tensor leaf_tensor =
+      egr_utils_api::CreateTensorWithValue(ddim,
+                                           paddle::platform::CPUPlace(),
+                                           phi::DataType::FLOAT32,
+                                           phi::DataLayout::NCHW,
+                                           1.0 /*value*/,
+                                           true /*is_leaf*/);
   {
     // Create Node0
     auto node0_ptr = std::make_shared<GradNodeScale>(1, 1);
@@ -220,10 +247,10 @@ TEST(Backward, LinearNodes) {
   }
 
   // Use Empty Grad Tensor
-  Backward(target_tensors, {});
+  auto result = Grad(target_tensors, {leaf_tensor}, {});
 
   // Check Output Value
-  eager_test::CompareGradTensorWithValue<float>(leaf_tensor, 50.0);
+  eager_test::CompareTensorWithValue<float>(result[0], 50.0);
 }
 
 /*
@@ -231,9 +258,9 @@ TEST(Backward, LinearNodes) {
     |   |
 Node0   Node1
   |      |
- inp0   inp1
+ in0   in1
 */
-TEST(Backward, WithAccumulation) {
+TEST(Grad, WithAccumulation) {
   // Prepare Device Contexts
   eager_test::InitEnv(paddle::platform::CPUPlace());
 
@@ -318,7 +345,7 @@ TEST(Backward, WithAccumulation) {
 
     // Connect Node1 -> Node2 via Edge
     auto tmp_tensor1 = paddle::Tensor();
-    auto* meta1 = EagerUtils::autograd_meta(&tmp_tensor1);
+    auto meta1 = EagerUtils::autograd_meta(&tmp_tensor1);
     meta1->SetStopGradient(false);
     meta1->SetSingleOutRankWithSlot(0, 0);
     meta1->SetGradNode(node2_ptr);
@@ -334,13 +361,12 @@ TEST(Backward, WithAccumulation) {
     auto_grad_meta2->SetSingleOutRankWithSlot(0, 0);
 
     auto_grad_meta2->SetStopGradient(false);
-    std::vector<egr::AutogradMeta*> res2 = {auto_grad_meta2};
     node2_ptr->SetGradOutMeta(leaf_tensor, 0);
   }
 
-  Backward(target_tensors, grad_tensors);
+  auto result = Grad(target_tensors, {leaf_tensor}, grad_tensors);
 
-  eager_test::CompareGradTensorWithValue<float>(leaf_tensor, 2500.0);
+  eager_test::CompareTensorWithValue<float>(result[0], 2500.0);
 }
 
 }  // namespace egr
