+++ test/cpp/eager/task_tests/hook_test_intermidiate.cc
@@ -20,7 +20,7 @@
 #include "paddle/fluid/eager/backward.h"
 #include "paddle/fluid/eager/grad_node_info.h"
 #include "paddle/fluid/eager/hooks.h"
-#include "paddle/fluid/eager/tests/test_utils.h"
+#include "paddle/fluid/eager/test_utils.h"
 #include "paddle/fluid/imperative/tracer.h"
 #include "paddle/phi/core/dense_tensor.h"
 #include "paddle/phi/core/kernel_registry.h"
