+++ test/cpp/eager/task_tests/tensor_utils_test.cc
@@ -20,7 +20,7 @@
 #include "paddle/fluid/eager/eager_tensor.h"
 #include "paddle/fluid/eager/grad_node_info.h"
 #include "paddle/fluid/eager/grad_tensor_holder.h"
-#include "paddle/fluid/eager/tests/test_utils.h"
+#include "paddle/fluid/eager/test_utils.h"
 #include "paddle/phi/api/lib/utils/allocator.h"
 #include "paddle/phi/core/kernel_registry.h"
